fix: keep browser selector sandbox compatible
